COMMERCE-5657 dispatch-service - Reorganize - rename package
